rust: If a property function produces a ParserError return that error
If a property function produces a ParserError that means the function
was found but produced an error, so assume that error is more correct
than anything else we can do here.

This ensures that when certain errors happen, we report them more
correctly instead of just "could not parse input".