rust: Correct order of parts in Selectors
As it turns out, selectors::Selector::iter_raw_parse_order does not
actually match the order of declaration in the CSS file. Handle this
more correctly, so that things like "&:hovered" get parsed as
"RelativeParent Class" instead of "Class RelativeParent", which is
important when we resolve the relative parent since we replace the
RelativeParent with the parent selector.